refactor feedback and point
