+++ app/usecase/feedback_point_repository.go
@@ -1,0 +1,7 @@
+package usecase
+
+import "project-symi-backend/app/domain"
+
+type FeedbackPointsRepository interface {
+	StoreFeedbackAndUpdatePoints(domain.StoredFeedback, string) (int, error)
+}
